+++ src/models/Stat.java
@@ -3,12 +3,16 @@ package models;
 // represents a stat (unused)
 
 public class Stat {
-	StatName name;
-	int val;
+	private StatName name;
+	private int val;
 	
 	public Stat(StatName _name, int _val) {
 		name = _name;
 		val = _val;
+	}
+	
+	public int getValue() {
+		return val;
 	}
 	
 }